update rust utilities
