+++ bellman_circuits/Cargo.toml
@@ -24,7 +24,8 @@ sha2 = "0.10.6"
 bls12_381 = "0.8.0"
 hex = "0.4.3"
 clap = { version = "4.2.7", features = ["derive"] }
-utilities = { path = "utilities" }
+bellman_utils = { path = "bellman_utils" }
+rust_utils = {path = "../rust_utils"}
 bincode = "1.3"
 serde = { version = "1.0.163", features = ["derive"] }
 serde_json = "1.0.96"
