+++ bellman_circuits/benches/benchmark_circuit.rs
@@ -11,7 +11,7 @@ use criterion::measurement::Measurement;
 use sha2::{Digest, Sha256};
 use bellman::gadgets::multipack;
 use bls12_381::{Bls12, Scalar};
-use utilities::{read_file_from_env_var, read_env_variable};
+use rust_utils::{read_file_from_env_var, read_env_variable};
 use ff::PrimeField;
 use bellman::gadgets::test::TestConstraintSystem;
