+++ rust_utils/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "utilities"
+name = "rust_utils"
 version = "0.1.0"
 edition = "2021"
