+++ rust_utils/src/lib.rs
@@ -3,9 +3,6 @@ use std::{env, fs::File};
 use std::io::Read;
 use std::process;
 use psutil;
-use std::fs;
-use bellman::groth16::Proof;
-use bls12_381::Bls12;
 
 #[derive(Serialize)]
 struct Results {
@@ -15,8 +12,6 @@ struct Results {
     verify_rss: u64,
     proof_size: usize,
 }
-
-
 
 pub fn read_file_contents(file_name: String) -> String {
     let mut file = File::open(file_name).expect("Cannot load file");
@@ -74,24 +69,4 @@ pub fn save_results(
         output_file_str,
         serialized
     ).expect("Could not write to file");
-}
-
-pub fn measure_size_in_bytes(proof: &Proof<Bls12>) -> usize {
-    // TODO: Should we serialize the proof in another format?
-    // Serialize data and save it to a temporary file
-    let temp_file_path = "temp_file.bin";
-
-    let mut file = File::create(temp_file_path).expect("Could not create temp file");
-    proof.write(&mut file).expect("Could not write proof to file");
-
-    // Measure the size of the file
-    let file_size: usize = fs::metadata(&temp_file_path).expect("Cannot read the size of temp file").len() as usize;
-
-    // Convert file size to MB
-    let size_in_mb = file_size;
-
-    // Remove the temporary file
-    fs::remove_file(&temp_file_path).expect("Cannot remove temp file");
-
-    return size_in_mb;
 }
